requirements.txt
aiohttp==1.0.5
async-timeout==1.3.0
discord.py==0.16.11
websockets==3.4
dotobject==1.2.1
redis==2.10.6
redisworks==0.2.7
